+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to pHTML JSX
 
+### 2.1.0 (February 20, 2019)
+
+- Updated: Allow Babel overrides for transform options and plugins.
+- Updated: Maintain line-accurate source mappings.
+
 ### 2.0.0 (February 18, 2019)
 
 - Updated: `phtml` to 2.0.1 (major)